+++ Heimdall.Server/Utilities/BifrostSubscription.cs
@@ -3,7 +3,8 @@
 namespace Heimdall.Server.Utilities
 {
     internal readonly record struct BifrostSubscription(
-    Guid Id,
-    ChannelReader<BifrostMessage> Reader,
-    Action Unsubscribe);
+        Guid Id,
+        ChannelReader<BifrostMessage> Reader,
+        Action Unsubscribe,
+        Task<string> DisconnectRequested);
 }